+++ include/flatcc/reflection/reflection_builder.h
@@ -21,36 +21,43 @@ __flatbuffers_build_scalar(flatbuffers_, reflection_BaseType, reflection_BaseTyp
 
 static const flatbuffers_voffset_t __reflection_Type_required[] = { 0 };
 __flatbuffers_build_table(flatbuffers_, reflection_Type, 3)
+static const flatbuffers_voffset_t __reflection_KeyValue_required[] = { 0, 0 };
+__flatbuffers_build_table(flatbuffers_, reflection_KeyValue, 2)
 static const flatbuffers_voffset_t __reflection_EnumVal_required[] = { 0, 0 };
 __flatbuffers_build_table(flatbuffers_, reflection_EnumVal, 4)
 static const flatbuffers_voffset_t __reflection_Enum_required[] = { 0, 1, 3, 0 };
-__flatbuffers_build_table(flatbuffers_, reflection_Enum, 4)
+__flatbuffers_build_table(flatbuffers_, reflection_Enum, 5)
 static const flatbuffers_voffset_t __reflection_Field_required[] = { 0, 1, 0 };
-__flatbuffers_build_table(flatbuffers_, reflection_Field, 9)
+__flatbuffers_build_table(flatbuffers_, reflection_Field, 10)
 static const flatbuffers_voffset_t __reflection_Object_required[] = { 0, 1, 0 };
-__flatbuffers_build_table(flatbuffers_, reflection_Object, 5)
+__flatbuffers_build_table(flatbuffers_, reflection_Object, 6)
 static const flatbuffers_voffset_t __reflection_Schema_required[] = { 0, 1, 0 };
 __flatbuffers_build_table(flatbuffers_, reflection_Schema, 5)
 #define __reflection_Type_formal_args , reflection_BaseType_enum_t v0, reflection_BaseType_enum_t v1, int32_t v2
 #define __reflection_Type_call_args , v0, v1, v2
 static inline reflection_Type_ref_t reflection_Type_create(flatbuffers_builder_t *B __reflection_Type_formal_args);
+#define __reflection_KeyValue_formal_args , flatbuffers_string_ref_t v0, flatbuffers_string_ref_t v1
+#define __reflection_KeyValue_call_args , v0, v1
+static inline reflection_KeyValue_ref_t reflection_KeyValue_create(flatbuffers_builder_t *B __reflection_KeyValue_formal_args);
 #define __reflection_EnumVal_formal_args , flatbuffers_string_ref_t v0, int64_t v1, reflection_Object_ref_t v2, reflection_Type_ref_t v3
 #define __reflection_EnumVal_call_args , v0, v1, v2, v3
 static inline reflection_EnumVal_ref_t reflection_EnumVal_create(flatbuffers_builder_t *B __reflection_EnumVal_formal_args);
-#define __reflection_Enum_formal_args , flatbuffers_string_ref_t v0, reflection_EnumVal_vec_ref_t v1, flatbuffers_bool_t v2, reflection_Type_ref_t v3
-#define __reflection_Enum_call_args , v0, v1, v2, v3
+#define __reflection_Enum_formal_args ,\
+  flatbuffers_string_ref_t v0, reflection_EnumVal_vec_ref_t v1, flatbuffers_bool_t v2, reflection_Type_ref_t v3, reflection_KeyValue_vec_ref_t v4
+#define __reflection_Enum_call_args ,\
+  v0, v1, v2, v3, v4
 static inline reflection_Enum_ref_t reflection_Enum_create(flatbuffers_builder_t *B __reflection_Enum_formal_args);
 #define __reflection_Field_formal_args ,\
   flatbuffers_string_ref_t v0, reflection_Type_ref_t v1, uint16_t v2, uint16_t v3,\
-  int64_t v4, double v5, flatbuffers_bool_t v6, flatbuffers_bool_t v7, flatbuffers_bool_t v8
+  int64_t v4, double v5, flatbuffers_bool_t v6, flatbuffers_bool_t v7, flatbuffers_bool_t v8, reflection_KeyValue_vec_ref_t v9
 #define __reflection_Field_call_args ,\
   v0, v1, v2, v3,\
-  v4, v5, v6, v7, v8
+  v4, v5, v6, v7, v8, v9
 static inline reflection_Field_ref_t reflection_Field_create(flatbuffers_builder_t *B __reflection_Field_formal_args);
 #define __reflection_Object_formal_args ,\
-  flatbuffers_string_ref_t v0, reflection_Field_vec_ref_t v1, flatbuffers_bool_t v2, int32_t v3, int32_t v4
+  flatbuffers_string_ref_t v0, reflection_Field_vec_ref_t v1, flatbuffers_bool_t v2, int32_t v3, int32_t v4, reflection_KeyValue_vec_ref_t v5
 #define __reflection_Object_call_args ,\
-  v0, v1, v2, v3, v4
+  v0, v1, v2, v3, v4, v5
 static inline reflection_Object_ref_t reflection_Object_create(flatbuffers_builder_t *B __reflection_Object_formal_args);
 #define __reflection_Schema_formal_args ,\
   reflection_Object_vec_ref_t v0, reflection_Enum_vec_ref_t v1, flatbuffers_string_ref_t v2, flatbuffers_string_ref_t v3, reflection_Object_ref_t v4
@@ -73,6 +80,20 @@ static inline reflection_Type_ref_t reflection_Type_create(flatbuffers_builder_t
     return reflection_Type_end(B);
 }
 __flatbuffers_build_table_prolog(flatbuffers_, reflection_Type, reflection_Type_identifier, reflection_Type_type_identifier)
+
+__flatbuffers_build_string_field(0, flatbuffers_, reflection_KeyValue_key)
+__flatbuffers_build_string_field(1, flatbuffers_, reflection_KeyValue_value)
+
+static inline reflection_KeyValue_ref_t reflection_KeyValue_create(flatbuffers_builder_t *B __reflection_KeyValue_formal_args)
+{
+    if (reflection_KeyValue_start(B)
+        || reflection_KeyValue_key_add(B, v0)
+        || reflection_KeyValue_value_add(B, v1)) {
+        return 0;
+    }
+    return reflection_KeyValue_end(B);
+}
+__flatbuffers_build_table_prolog(flatbuffers_, reflection_KeyValue, reflection_KeyValue_identifier, reflection_KeyValue_type_identifier)
 
 __flatbuffers_build_string_field(0, flatbuffers_, reflection_EnumVal_name)
 __flatbuffers_build_scalar_field(1, flatbuffers_, reflection_EnumVal_value, flatbuffers_int64, int64_t, 8, 8, INT64_C(0))
@@ -97,6 +118,8 @@ __flatbuffers_build_string_field(0, flatbuffers_, reflection_Enum_name)
 __flatbuffers_build_table_vector_field(1, flatbuffers_, reflection_Enum_values, reflection_EnumVal)
 __flatbuffers_build_scalar_field(2, flatbuffers_, reflection_Enum_is_union, flatbuffers_bool, flatbuffers_bool_t, 1, 1, UINT8_C(0))
 __flatbuffers_build_table_field(3, flatbuffers_, reflection_Enum_underlying_type, reflection_Type)
+/* vector has keyed elements */
+__flatbuffers_build_table_vector_field(4, flatbuffers_, reflection_Enum_attributes, reflection_KeyValue)
 
 static inline reflection_Enum_ref_t reflection_Enum_create(flatbuffers_builder_t *B __reflection_Enum_formal_args)
 {
@@ -104,6 +127,7 @@ static inline reflection_Enum_ref_t reflection_Enum_create(flatbuffers_builder_t
         || reflection_Enum_name_add(B, v0)
         || reflection_Enum_values_add(B, v1)
         || reflection_Enum_underlying_type_add(B, v3)
+        || reflection_Enum_attributes_add(B, v4)
         || reflection_Enum_is_union_add(B, v2)) {
         return 0;
     }
@@ -120,6 +144,8 @@ __flatbuffers_build_scalar_field(5, flatbuffers_, reflection_Field_default_real,
 __flatbuffers_build_scalar_field(6, flatbuffers_, reflection_Field_deprecated, flatbuffers_bool, flatbuffers_bool_t, 1, 1, UINT8_C(0))
 __flatbuffers_build_scalar_field(7, flatbuffers_, reflection_Field_required, flatbuffers_bool, flatbuffers_bool_t, 1, 1, UINT8_C(0))
 __flatbuffers_build_scalar_field(8, flatbuffers_, reflection_Field_key, flatbuffers_bool, flatbuffers_bool_t, 1, 1, UINT8_C(0))
+/* vector has keyed elements */
+__flatbuffers_build_table_vector_field(9, flatbuffers_, reflection_Field_attributes, reflection_KeyValue)
 
 static inline reflection_Field_ref_t reflection_Field_create(flatbuffers_builder_t *B __reflection_Field_formal_args)
 {
@@ -128,6 +154,7 @@ static inline reflection_Field_ref_t reflection_Field_create(flatbuffers_builder
         || reflection_Field_default_real_add(B, v5)
         || reflection_Field_name_add(B, v0)
         || reflection_Field_type_add(B, v1)
+        || reflection_Field_attributes_add(B, v9)
         || reflection_Field_id_add(B, v2)
         || reflection_Field_offset_add(B, v3)
         || reflection_Field_deprecated_add(B, v6)
@@ -145,6 +172,8 @@ __flatbuffers_build_table_vector_field(1, flatbuffers_, reflection_Object_fields
 __flatbuffers_build_scalar_field(2, flatbuffers_, reflection_Object_is_struct, flatbuffers_bool, flatbuffers_bool_t, 1, 1, UINT8_C(0))
 __flatbuffers_build_scalar_field(3, flatbuffers_, reflection_Object_minalign, flatbuffers_int32, int32_t, 4, 4, INT32_C(0))
 __flatbuffers_build_scalar_field(4, flatbuffers_, reflection_Object_bytesize, flatbuffers_int32, int32_t, 4, 4, INT32_C(0))
+/* vector has keyed elements */
+__flatbuffers_build_table_vector_field(5, flatbuffers_, reflection_Object_attributes, reflection_KeyValue)
 
 static inline reflection_Object_ref_t reflection_Object_create(flatbuffers_builder_t *B __reflection_Object_formal_args)
 {
@@ -153,6 +182,7 @@ static inline reflection_Object_ref_t reflection_Object_create(flatbuffers_build
         || reflection_Object_fields_add(B, v1)
         || reflection_Object_minalign_add(B, v3)
         || reflection_Object_bytesize_add(B, v4)
+        || reflection_Object_attributes_add(B, v5)
         || reflection_Object_is_struct_add(B, v2)) {
         return 0;
     }