+++ include/flatcc/reflection/reflection_verifier.h
@@ -10,6 +10,7 @@
 #include "flatcc/flatcc_prologue.h"
 
 static int reflection_Type_verify_table(flatcc_table_verifier_descriptor_t *td);
+static int reflection_KeyValue_verify_table(flatcc_table_verifier_descriptor_t *td);
 static int reflection_EnumVal_verify_table(flatcc_table_verifier_descriptor_t *td);
 static int reflection_Enum_verify_table(flatcc_table_verifier_descriptor_t *td);
 static int reflection_Field_verify_table(flatcc_table_verifier_descriptor_t *td);
@@ -43,6 +44,34 @@ static inline int reflection_Type_verify_as_root_with_identifier(const void *buf
 static inline int reflection_Type_verify_as_root_with_type_hash(const void *buf, size_t bufsiz, flatbuffers_thash_t thash)
 {
     return flatcc_verify_table_as_typed_root(buf, bufsiz, thash, &reflection_Type_verify_table);
+}
+
+static int reflection_KeyValue_verify_table(flatcc_table_verifier_descriptor_t *td)
+{
+    int ret;
+    if ((ret = flatcc_verify_string_field(td, 0, 1) /* key */)) return ret;
+    if ((ret = flatcc_verify_string_field(td, 1, 0) /* value */)) return ret;
+    return flatcc_verify_ok;
+}
+
+static inline int reflection_KeyValue_verify_as_root(const void *buf, size_t bufsiz)
+{
+    return flatcc_verify_table_as_root(buf, bufsiz, reflection_KeyValue_identifier, &reflection_KeyValue_verify_table);
+}
+
+static inline int reflection_KeyValue_verify_as_typed_root(const void *buf, size_t bufsiz)
+{
+    return flatcc_verify_table_as_root(buf, bufsiz, reflection_KeyValue_type_identifier, &reflection_KeyValue_verify_table);
+}
+
+static inline int reflection_KeyValue_verify_as_root_with_identifier(const void *buf, size_t bufsiz, const char *fid)
+{
+    return flatcc_verify_table_as_root(buf, bufsiz, fid, &reflection_KeyValue_verify_table);
+}
+
+static inline int reflection_KeyValue_verify_as_root_with_type_hash(const void *buf, size_t bufsiz, flatbuffers_thash_t thash)
+{
+    return flatcc_verify_table_as_typed_root(buf, bufsiz, thash, &reflection_KeyValue_verify_table);
 }
 
 static int reflection_EnumVal_verify_table(flatcc_table_verifier_descriptor_t *td)
@@ -82,6 +111,7 @@ static int reflection_Enum_verify_table(flatcc_table_verifier_descriptor_t *td)
     if ((ret = flatcc_verify_table_vector_field(td, 1, 1, &reflection_EnumVal_verify_table) /* values */)) return ret;
     if ((ret = flatcc_verify_field(td, 2, 1, 1) /* is_union */)) return ret;
     if ((ret = flatcc_verify_table_field(td, 3, 1, &reflection_Type_verify_table) /* underlying_type */)) return ret;
+    if ((ret = flatcc_verify_table_vector_field(td, 4, 0, &reflection_KeyValue_verify_table) /* attributes */)) return ret;
     return flatcc_verify_ok;
 }
 
@@ -117,6 +147,7 @@ static int reflection_Field_verify_table(flatcc_table_verifier_descriptor_t *td)
     if ((ret = flatcc_verify_field(td, 6, 1, 1) /* deprecated */)) return ret;
     if ((ret = flatcc_verify_field(td, 7, 1, 1) /* required */)) return ret;
     if ((ret = flatcc_verify_field(td, 8, 1, 1) /* key */)) return ret;
+    if ((ret = flatcc_verify_table_vector_field(td, 9, 0, &reflection_KeyValue_verify_table) /* attributes */)) return ret;
     return flatcc_verify_ok;
 }
 
@@ -148,6 +179,7 @@ static int reflection_Object_verify_table(flatcc_table_verifier_descriptor_t *td
     if ((ret = flatcc_verify_field(td, 2, 1, 1) /* is_struct */)) return ret;
     if ((ret = flatcc_verify_field(td, 3, 4, 4) /* minalign */)) return ret;
     if ((ret = flatcc_verify_field(td, 4, 4, 4) /* bytesize */)) return ret;
+    if ((ret = flatcc_verify_table_vector_field(td, 5, 0, &reflection_KeyValue_verify_table) /* attributes */)) return ret;
     return flatcc_verify_ok;
 }